requirements.txt
# Cross-lingual Prompt Consistency Evaluation Dependencies

# LLM inference
ollama>=0.1.0

# Embeddings and NLP
sentence-transformers>=2.2.0
torch>=2.0.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning utilities
scikit-learn>=1.3.0

# Notebooks
jupyter>=1.0.0
ipywidgets>=8.0.0

# Configuration
pyyaml>=6.0.0

# Utilities
tqdm>=4.65.0
